// src/app/interfaces/Movie.ts
// Generated by https://quicktype.io

import { Personal } from "./Anime";

export interface MovieResult {
    page:          number;
    results:       Movie[];
    total_pages:   number;
    total_results: number;
}

export interface Movie {
    adult:             boolean;
    backdrop_path:     string;
    genre_ids:         number[];
    id:                number;
    original_language: OriginalLanguage;
    original_title:    string;
    overview:          string;
    popularity:        number;
    poster_path:       string;
    release_date:      Date;
    title:             string;
    video:             boolean;
    vote_average:      number;
    vote_count:        number;
    isSaved?: boolean;
    personal?: Personal;
}

export enum OriginalLanguage {
    De = "de",
    En = "en",
    Hi = "hi",
    Ja = "ja",
}

export interface GenreMovies {
    genres: Genre[];
}

export interface Genre {
    id:   number;
    name: string;
}